Template Organization - Moved template files to a dedicated folder
- Created a 'templates' folder to store templates like footer, navigation, and abstract template for better organization.
- Fixed import paths to align with the new folder structure and resolved any related issues.
