+++ js/template/templates/AbstractTemplate.js
