+++ js/template/templates/footer.js
