+++ js/template/templates/footer/acknowledge.js
