+++ js/template/templates/footer/copyright.js
