+++ js/template/templates/footer/navigations.js
